+++ package/mine.cc
@@ -15,8 +15,8 @@
 //
 #include <algorithm>
 #include <random>
-#include <iterator>
 #include <sstream>
+#include <iterator>
 #include <stdexcept>
 #include <iostream>
 #include <vector>
@@ -89,6 +89,22 @@ ByteArray MineCommon::generateRandomBytes(const std::size_t len) noexcept
     return result;
 }
 
+std::string MineCommon::byteArrayToRawString(const ByteArray& input) noexcept
+{
+    std::ostringstream ss;
+    std::copy(input.begin(), input.end(), std::ostream_iterator<char>(ss));
+    return ss.str();
+}
+
+ByteArray MineCommon::rawStringToByteArray(const std::string& str) noexcept
+{
+    ByteArray byteArr;
+    for (char c : str) {
+        byteArr.push_back(static_cast<byte>(c));
+    }
+    return byteArr;
+}
+
 std::string MineCommon::version() noexcept
 {
     return MINE_VERSION;
@@ -118,13 +134,6 @@ ByteArray Base16::fromString(const std::string& hex)
         byteArr.push_back(encode(hex.substr(i, 2).c_str()));
     }
     return byteArr;
-}
-
-std::string Base16::toRawString(const ByteArray& input)
-{
-    std::ostringstream ss;
-    std::copy(input.begin(), input.end(), std::ostream_iterator<char>(ss));
-    return ss.str();
 }
 
 void Base16::decode(char a, char b, std::ostringstream& ss)
@@ -761,7 +770,7 @@ ByteArray AES::resolveInputMode(const std::string& input, MineCommon::Encoding i
 std::string AES::resolveOutputMode(const ByteArray& input, MineCommon::Encoding outputMode)
 {
     if (outputMode == MineCommon::Encoding::Raw) {
-        return Base16::toRawString(input);
+        return MineCommon::byteArrayToRawString(input);
     } else if (outputMode == MineCommon::Encoding::Base16) {
         return Base16::encode(input.begin(), input.end());
     }
@@ -1039,7 +1048,7 @@ std::string AES::encr(const std::string& input, MineCommon::Encoding inputEncodi
     if (m_key.empty()) {
         throw std::runtime_error("Key not set");
     }
-    return encrypt(input, Base16::encode(Base16::toRawString(m_key)), inputEncoding, outputEncoding, pkcs5Padding);
+    return encrypt(input, Base16::encode(MineCommon::byteArrayToRawString(m_key)), inputEncoding, outputEncoding, pkcs5Padding);
 }
 
 std::string AES::encr(const std::string& input, std::string& iv, MineCommon::Encoding inputEncoding, MineCommon::Encoding outputEncoding, bool pkcs5Padding)
@@ -1047,7 +1056,7 @@ std::string AES::encr(const std::string& input, std::string& iv, MineCommon::Enc
     if (m_key.empty()) {
         throw std::runtime_error("Key not set");
     }
-    return encrypt(input, Base16::encode(Base16::toRawString(m_key)), iv, inputEncoding, outputEncoding, pkcs5Padding);
+    return encrypt(input, Base16::encode(MineCommon::byteArrayToRawString(m_key)), iv, inputEncoding, outputEncoding, pkcs5Padding);
 }
 
 std::string AES::decr(const std::string& input, MineCommon::Encoding inputEncoding, MineCommon::Encoding outputEncoding)
@@ -1055,7 +1064,7 @@ std::string AES::decr(const std::string& input, MineCommon::Encoding inputEncodi
     if (m_key.empty()) {
         throw std::runtime_error("Key not set");
     }
-    return decrypt(input, Base16::encode(Base16::toRawString(m_key)), inputEncoding, outputEncoding);
+    return decrypt(input, Base16::encode(MineCommon::byteArrayToRawString(m_key)), inputEncoding, outputEncoding);
 }
 
 std::string AES::decr(const std::string& input, const std::string& iv, MineCommon::Encoding inputEncoding, MineCommon::Encoding outputEncoding)
@@ -1063,7 +1072,7 @@ std::string AES::decr(const std::string& input, const std::string& iv, MineCommo
     if (m_key.empty()) {
         throw std::runtime_error("Key not set");
     }
-    return decrypt(input, Base16::encode(Base16::toRawString(m_key)), iv, inputEncoding, outputEncoding);
+    return decrypt(input, Base16::encode(MineCommon::byteArrayToRawString(m_key)), iv, inputEncoding, outputEncoding);
 }
 
 ByteArray AES::encr(const ByteArray& input, bool pkcs5Padding)