+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [Unreleased]
+### Updates
+- Moved `Base16::toRawString` to `MineCommon::byteArrayToRawString`
+- Added `MineCommon::rawStringToByteArray`
+
 ## [1.0.0] - 05-09-2017
 ### Added
 - Initial release